feat: update file structure and add html templates
- add home page template
