+++ ui/html/home.page.tmpl
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Snippet bin</title>
+</head>
+<body>
+    <header>
+        <h1><a href="/">Snippet-bin</a></h1>
+    </header>
+    <nav>
+        <a href="/">Home</a>
+    </nav>
+    <main>
+        <h2>Latest Snippets</h2>
+        <p>There's nothing to see here yet</p>
+    </main>
+</body>
+</html>
